pyproject.toml
[project]
name = "lead-enrichment"
version = "0.1.0"
description = "B2B lead-form enrichment"
authors = [{ name = "Karsten Bjerring", email = "karsols@dfds.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core HTTP
    "requests>=2.32.4",
    "urllib3>=2.5.0",
    # Data processing
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "duckdb>=1.3.2",
    # Web scraping
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "justext>=3.0.2",
    "readability-lxml>=0.8.4.1",
    # AI/LLM
    "openai>=1.108.0",
    "openai-agents>=0.3.3",
    # Data validation
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    # Authentication
    "azure-identity>=1.23.1",
    "msal>=1.33.0",
    # Utilities
    "googletrans>=4.0.2",
    "loguru>=0.7.3",
    "ddgs>=9.6.0",
    # Development
    "jupyter>=1.1.1",
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.30.0",
]

[project.scripts]
dev = "scripts.dev:main"
start = "scripts.prod:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

# Ruff (formatting + linting)
[tool.ruff]
line-length = 120
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]
ignore = ["E501"]

# Import sorting
[tool.ruff.isort]
known-first-party = ["common", "services", "tools", "custom_agents"]

# Type checking
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
strict_optional = true

# Tests
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["*_test.py", "test_*.py"]
python_classes = "Test*"
python_functions = "test_*"
addopts = ["--import-mode=importlib"]

[tool.uv]
package = true

[tool.uv.sources]
content-classifier = { workspace = true }
segment-public-api = { git = "https://github.com/segmentio/public-api-sdk-python.git" }

[dependency-groups]
dev = [
    # Testing
    "pytest>=8.4.1",
    "pytest-cases>=3.8.6",
    # Code quality
    "ruff>=0.12.1",
    "mypy>=1.7.0",
    "pre-commit>=4.2.0",
    # Type stubs
    "types-requests>=2.31.0",
    "types-pyyaml>=6.0.0",
    # Jupyter/notebooks
    "ipykernel>=6.29.5",
    "jupyterlab>=4.2.5",
]
